+++ declarations/autojs6/aj6-main-app.d.ts
@@ -8972,6 +8972,8 @@ declare namespace org {
 						public newMat(mat: __javaRoots.orgRoot.autojs.autojs.core.opencv.Mat, rect: __javaRoots.orgRoot.opencv.core.Rect): __javaRoots.orgRoot.autojs.autojs.core.opencv.Mat;
 						public static pixel(image: __javaRoots.orgRoot.autojs.autojs.core.image.ImageWrapper, x: number, y: number): number;
 						public quantize(image: __javaRoots.orgRoot.autojs.autojs.core.image.ImageWrapper, options: __javaRoots.orgRoot.autojs.autojs.runtime.api.PngQuantBridge.Options): __javaRoots.orgRoot.autojs.autojs.runtime.api.PngQuantBridge.Result;
+						public quantizeToFile(image: __javaRoots.orgRoot.autojs.autojs.core.image.ImageWrapper, path: string, options: __javaRoots.orgRoot.autojs.autojs.runtime.api.PngQuantBridge.Options): __javaRoots.orgRoot.autojs.autojs.runtime.api.PngQuantBridge.OutputResult;
+						public quantizeToFileDescriptor(image: __javaRoots.orgRoot.autojs.autojs.core.image.ImageWrapper, fileDescriptor: __javaRoots.javaRoot.io.FileDescriptor, options: __javaRoots.orgRoot.autojs.autojs.runtime.api.PngQuantBridge.Options): __javaRoots.orgRoot.autojs.autojs.runtime.api.PngQuantBridge.OutputResult;
 						public read(path: string): __javaRoots.orgRoot.autojs.autojs.core.image.ImageWrapper;
 						public read(path: string, isStrict: boolean): __javaRoots.orgRoot.autojs.autojs.core.image.ImageWrapper;
 						public releaseScreenCapturer(): void;
@@ -11789,6 +11791,7 @@ declare namespace org {
 						public static probeLoadedRuntime(): string;
 						public static quantize(bitmap: __javaRoots.androidRoot.graphics.Bitmap, quality: number): number[];
 						public static quantize(bitmap: __javaRoots.androidRoot.graphics.Bitmap, options: __javaRoots.orgRoot.autojs.autojs.runtime.api.PngQuantBridge.Options): number[];
+						public static quantizeToFileDescriptor(bitmap: __javaRoots.androidRoot.graphics.Bitmap, fileDescriptor: __javaRoots.javaRoot.io.FileDescriptor, options: __javaRoots.orgRoot.autojs.autojs.runtime.api.PngQuantBridge.Options): __javaRoots.orgRoot.autojs.autojs.runtime.api.PngQuantBridge.OutputResult;
 						public static quantizeWithResult(bitmap: __javaRoots.androidRoot.graphics.Bitmap, options: __javaRoots.orgRoot.autojs.autojs.runtime.api.PngQuantBridge.Options): __javaRoots.orgRoot.autojs.autojs.runtime.api.PngQuantBridge.Result;
 					}
 					export namespace PngQuantBridge {
@@ -11797,20 +11800,33 @@ declare namespace org {
 							public static readonly DEFAULT_MAX_COLORS: number;
 							public static readonly DEFAULT_MIN_QUALITY: number;
 							public static readonly DEFAULT_POSTERIZE_BITS: number;
+							public static readonly DEFAULT_PRESERVE_ALPHA: boolean;
 							public static readonly DEFAULT_SPEED: number;
 							public readonly ditheringLevel: number;
 							public readonly maxColors: number;
 							public readonly maxQuality: number;
 							public readonly minQuality: number;
 							public readonly posterizeBits: number;
+							public readonly preserveAlpha: boolean;
 							public readonly speed: number;
 							public constructor(maxColors: number, speed: number, minQuality: number, maxQuality: number, ditheringLevel: number, posterizeBits: number);
+							public constructor(maxColors: number, speed: number, minQuality: number, maxQuality: number, ditheringLevel: number, posterizeBits: number, preserveAlpha: boolean);
 							public getDitheringLevel(): number;
 							public getMaxColors(): number;
 							public getMaxQuality(): number;
 							public getMinQuality(): number;
 							public getPosterizeBits(): number;
 							public getSpeed(): number;
+							public isPreserveAlpha(): boolean;
+						}
+						export class OutputResult {
+							public readonly quality: number;
+							public readonly quantizationError: number;
+							public readonly size: number;
+							public constructor(size: number, quality: number, quantizationError: number);
+							public getQuality(): number;
+							public getQuantizationError(): number;
+							public getSize(): number;
 						}
 						export class QualityTooLowException extends __javaRoots.javaRoot.io.IOException {
 							public static readonly CODE: string;